requirements.txt
django
pyjuggling
django-debug-toolbar
django-extensions
django-bootstrap3